Crude impl of summmary/message fetch
svn path=/branches/EXCHANGE_MAPI_BRANCH/; revision=8204

+++ servers/mapi/exchange-mapi-folder.h
@@ -47,7 +47,7 @@ exchange_mapi_folder_new (const char *folder_name, const char *parent_folder_nam
 			  uint32_t child_count, uint32_t unread_count, uint32_t total);
 
 const gchar* exchange_mapi_folder_get_name (ExchangeMAPIFolder *folder);
-const guint32 exchange_mapi_folder_get_fid (ExchangeMAPIFolder *folder);
+const guint64 exchange_mapi_folder_get_fid (ExchangeMAPIFolder *folder);
 const guint64 exchange_mapi_folder_get_parent_id (ExchangeMAPIFolder *folder);
 const guint32 exchange_mapi_folder_get_unread_count (ExchangeMAPIFolder *folder);
 const guint32 exchange_mapi_folder_get_total_count (ExchangeMAPIFolder *folder);
